# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CCAT Practice App — Gateway API
  version: "1.0.0-phase0"
  description: >
    Normative /v1 API for the Concept Mastery CCAT Practice App. The Gateway is the ONLY
    application-data boundary exposed to the student app (Blueprint v10.0 §1.1, §33).
    Direct student Supabase/PostgREST access is prohibited (§41). All sensitive mutations
    use idempotency keys; all mutable Admin resources use expected-version/ETag concurrency
    (§22.4, §32.1). This document is Phase 0 and covers the endpoint surface of §32; request
    bodies for Admin write operations are represented by shared envelopes and will be
    expanded per-domain in Phase 1.
  contact:
    name: Concept Mastery
    email: admin@conceptmastery.ca
servers:
  - url: https://api.conceptmastery.ca/v1
    description: Production
  - url: https://staging-api.conceptmastery.ca/v1
    description: Staging (no real child PII, §38.1)
...
security:
  - studentSession: []
paths:
...
  /recovery/pin/start:
    post:
      tags: [auth]
      operationId: recoveryPinStart
      security: []
      summary: Begin guardian-OTP PIN recovery (§4.4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, channel]
              properties:
                username: { type: string }
                channel: { $ref: "#/components/schemas/ContactChannel" }
      responses:
        "202": { $ref: "#/components/responses/ChallengeStarted" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /recovery/pin/complete:
    post:
      tags: [auth]
      operationId: recoveryPinComplete
      security: []
      summary: Verify OTP, set new PIN, revoke existing app sessions (§4.4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code, new_pin]
              properties:
                challenge_id: { type: string, format: uuid }
                code: { type: string }
                new_pin: { type: string, pattern: "^[0-9]{4}$" }
      responses:
        "200": { description: PIN reset; fresh login required }
        "401": { $ref: "#/components/responses/Error" }

  # ----------------------------------------------------------------- devices
  /devices/replacement/start:
    post:
      tags: [devices]
      operationId: deviceReplacementStart
      security: []
      summary: Begin single-device replacement (§5.2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, new_device_hash, channel]
              properties:
                username: { type: string }
                new_device_hash: { type: string }
                channel: { $ref: "#/components/schemas/ContactChannel" }
      responses:
        "202": { $ref: "#/components/responses/ChallengeStarted" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /devices/replacement/verify:
    post:
      tags: [devices]
      operationId: deviceReplacementVerify
      security: []
      summary: Verify OTP; revoke old device + sessions; enroll new sole device (§5.2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                challenge_id: { type: string, format: uuid }
                code: { type: string }
      responses:
        "200": { $ref: "#/components/responses/TokenPair" }
        "401": { $ref: "#/components/responses/Error" }
...
  /bookmarks:
    get:
      tags: [learning]
      operationId: listBookmarks
      responses:
        "200":
          description: Bookmarks
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/Bookmark" } }
    put:
      tags: [learning]
      operationId: putBookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [logical_question_id]
              properties:
                logical_question_id: { type: string, format: uuid }
                note: { type: string }
      responses:
        "200": { description: Bookmarked }
    delete:
      tags: [learning]
      operationId: deleteBookmark
      parameters:
        - name: logical_question_id
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Removed }

  # ----------------------------------------------------------------- rewards
...
components:
  securitySchemes:
    studentSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Student application session token, bound to the sole enrolled device (§5.4).
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin identity token (email/password + mandatory MFA, §22.1). Permissions loaded server-side.
...
  responses:
    Error:
      description: Structured error envelope (§32.1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Rate limited (§36.4). Fails closed.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ChallengeStarted:
      description: OTP challenge created
      content:
        application/json:
          schema:
            type: object
            properties:
              challenge_id: { type: string, format: uuid }
              expires_at: { type: string, format: date-time }
    TokenPair:
      description: Access + refresh tokens
      content:
        application/json:
          schema:
            type: object
            required: [access_token, refresh_token, expires_in]
            properties:
              access_token: { type: string }
              refresh_token: { type: string }
              expires_in: { type: integer }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Machine code, e.g. ACTIVE_SESSION_EXISTS }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    ContactChannel: { type: string, enum: [email, sms] }
...
    Bookmark:
      type: object
      properties:
        logical_question_id: { type: string, format: uuid }
        note: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
...